requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.2
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
Flask==3.0.0
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
requests==2.26.0
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.32
tzlocal==5.2
urllib3==1.26.18
Werkzeug==3.0.3
